feat: use flask-sqlalchemy-lite instead of flask-sqlalchemy + update database models
TODO: fix migrations
TODO: fix issues with the database
TODO: fix issues with some buttons that doesn't work anymore

This change introduce new mapped classes of SQLAlchemy, which are way better for typing purposes.

This commit is quite huge, while I am trying to keep it as concise as possible.