feat: don't return inside context manager
This is part of a big tentative to fix a bug, but the conclusion is that this bug is maybe not solvable.
Returning inside a context manager WILL end the context before continuing, so there is no issue here.
Still commiting because I made these little changes and now I keep them.