fix: Refactor OIDC authentication and improve database session management
- Ran `uv run tox` and fix the issues
- Updated `init_keycloak` to fetch well-known configuration conditionally based on the environment.
- Enhanced error handling in `oidc_info` to raise appropriate exceptions for missing claims.
- Added tests for the new authentication method `oidc_info`
- Improved database session management across various repositories by using `sessionmaker.begin()` for better transaction handling (that way the db is committed at the end). Also this supports nested sessions without issues.
- Added mock OIDC info for testing purposes to facilitate unit tests without relying on external services.
- Ensured that all database interactions in tests use context managers for session management.

I changed the JWT Bearer back to OAuth2 to force Connexion to send `required_context` to the authentication function. According to the OpenAPI documentation OAuth2 is the only protocol which supports these except OIDC (I think) but OIDC was a pain in the ass to get to work (I couldn't and I created an issue where I pinpoint the root cause). That is why I used OAuth2 but the underlying expectation of the functions are for the token to contain JWT data.
This can be changed to OpenID Connect when Connexion fixes their code.