feat: better MAC input fix:responsive search box and MAC inputs fix:valid payment for one day after creating account
Before: you create an account for a club, they can create a vm, the next day they can't access their vm because their payment expired, so you have to go back and validate the payment.
After: you create an account and don't forget to add a payment because they can't connect when they try it.